+++ src/components/homepage/navbar.js
@@ -24,9 +24,12 @@ export default function NavigationBar(props) {
           <SettingsButton
             className="button-settings"
             handleEditBg={(url) => props.handlePassBgUrl(url)}
+            handleCancelChanges={props.cancelChanges}
+            handleSaveChanges={props.saveChanges}
             location={props.location}
             bookmarkRef={props.bookmarkRef}
             setJiggle={props.setJiggle}
+            setFilter={props.setFilter}
           />
         </Grid>
       </Grid>